refactor(user): 重构用户模块代码
- 重构 UserRepo 中的 GetUserByUsername 和 CreateUser 方法，使用 UserDO 进行数据库操作
- 在 logout 服务中引入 mysql 包以支持数据库操作
- 更新 main.go，初始化 RPC 客户端
